+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Fix bug concerning `Exclude` properties inherited from a higher directory level. ([@jonas054][])
 * [#1500](https://github.com/bbatsov/rubocop/issues/1500): Fix crashing `--auto-correct --only IndentationWidth`. ([@jonas054][])
 * [#1512](https://github.com/bbatsov/rubocop/issues/1512): Fix false negative for typical string formatting examples. ([@kakutani][], [@jonas054][])
+* [#1504](https://github.com/bbatsov/rubocop/issues/1504): Fail with a meaningful error if the configuration file is malformed. ([@bquorning][])
 
 ## 0.28.0 (10/12/2014)